+++ src/main/java/com/leasecompany/carleasingsystem/database/data/inventoryItem/InventoryItemDAO.java
@@ -2,6 +2,6 @@ package com.leasecompany.carleasingsystem.database.data.inventoryItem;
 
 import com.leasecompany.carleasingsystem.database.data.GenericDAO;
 
-public interface InventoryItemDAO extends GenericDAO<InventoryItem, Integer> {
+public interface InventoryItemDAO extends GenericDAO<InventoryItem, Long> {
 
 }
