+++ src/main/java/com/leasecompany/carleasingsystem/database/data/inventoryItem/InventoryItemDAOImpl.java
@@ -3,7 +3,7 @@ package com.leasecompany.carleasingsystem.database.data.inventoryItem;
 import com.leasecompany.carleasingsystem.database.data.GenericDAOImpl;
 import org.hibernate.SessionFactory;
 
-public class InventoryItemDAOImpl extends GenericDAOImpl<InventoryItem, Integer> implements InventoryItemDAO {
+public class InventoryItemDAOImpl extends GenericDAOImpl<InventoryItem, Long> implements InventoryItemDAO {
 
     public InventoryItemDAOImpl(SessionFactory sessionFactory) {
         super(InventoryItem.class, sessionFactory);
